refactor: simplify GaussianBVH bounds structure and improve AABB computation
- Replaced std::pair with dedicated BVHBounds struct for better type safety
- Fixed data type consistency by using Vector3f instead of Vector3d
- Optimized AABB computation with improved memory layout and SIMD-friendly operations
- Enhanced numerical stability in covariance matrix calculations
- Added proper bounds validation and error handling for invalid Gaussians

+++ include/splat/spatial/gaussian_aabb.h
@@ -45,8 +45,8 @@ struct GaussianExtentsResult {
 
   /** Scene bounds (union of all Gaussian AABBs) */
   struct {
-    Eigen::Vector3d min;
-    Eigen::Vector3d max;
+    Eigen::Vector3f min;
+    Eigen::Vector3f max;
   } sceneBounds;
   /** Number of Gaussians skipped due to invalid values */
   size_t invalidCount{0};
